# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book API",
        "description": "A test for CSC Group",
        "contact": {
            "email": "lucas.mendes.dev@outlook.com.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Developmente Server"
        }
    ],
    "paths": {
        "v1/book": {
            "get": {
                "summary": "Retrieves a book pagination",
                "operationId": "a288eb04eeffd0abc792288e485b2dc3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of Book per page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Book Sort",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagination Result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/DefaultResponse"
                                        },
                                        {
                                            "properties": {
                                                "request": {
                                                    "example": "http://localhost/api/v1/book?search=test&sort=desc&page=1&limit=10"
                                                },
                                                "data": {
                                                    "$ref": "#/components/schemas/BookCollectionResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/DefaultResponse"
                                        },
                                        {
                                            "properties": {
                                                "code": {
                                                    "example": "500"
                                                },
                                                "method": {
                                                    "example": "GET"
                                                },
                                                "success": {
                                                    "example": "false"
                                                },
                                                "error": {
                                                    "example": "Internal Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a Book",
                "operationId": "524dc32fd3793da905462075c8b17bbb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/DefaultResponse"
                                        },
                                        {
                                            "properties": {
                                                "request": {
                                                    "example": "http://localhost/api/v1/book"
                                                },
                                                "data": {
                                                    "$ref": "#/components/schemas/ShowBookResource"
                                                },
                                                "code": {
                                                    "example": "201"
                                                },
                                                "method": {
                                                    "example": "POST"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/DefaultResponse"
                                        },
                                        {
                                            "properties": {
                                                "code": {
                                                    "example": "500"
                                                },
                                                "method": {
                                                    "example": "Delete"
                                                },
                                                "success": {
                                                    "example": "false"
                                                },
                                                "error": {
                                                    "example": "Internal Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateBookRequest": {
                "title": "Create Book",
                "properties": {
                    "title": {
                        "description": "Book title",
                        "type": "string",
                        "example": "Clean Arch"
                    },
                    "description": {
                        "description": "Book description",
                        "type": "string",
                        "example": "Book of Clean Arch"
                    },
                    "quantity": {
                        "description": "Book Quantity",
                        "type": "integer",
                        "example": "1"
                    },
                    "price": {
                        "description": "Book price",
                        "type": "integer",
                        "example": "100"
                    }
                },
                "type": "object"
            },
...
            "ShowAuthorResource": {
                "title": "Author Resource",
                "properties": {
                    "id": {
                        "description": "Author Uuid",
                        "type": "string",
                        "example": "7aa02c28-f2bd-387c-9c8c-b2069e0e6159"
                    },
                    "name": {
                        "description": "Author Name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "biography": {
                        "description": "Author Bio",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ArrayOfBooks": {
                "title": "Aray of Books",
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ShowBookResource"
                }
            },
            "BookCollectionResource": {
                "title": "Book Resource",
                "properties": {
                    "count": {
                        "description": "Page to retrieve",
                        "type": "integer"
                    },
                    "currentPage": {
                        "description": "Limit of Book per page",
                        "type": "integer"
                    },
                    "lastPage": {
                        "description": "Last Page",
                        "type": "integer"
                    },
                    "items": {
                        "$ref": "#/components/schemas/ArrayOfBooks"
                    }
                },
                "type": "object"
            },
            "ShowBookResource": {
                "title": "Book Resource",
                "properties": {
                    "id": {
                        "description": "Book Uuid",
                        "type": "string",
                        "example": "7aa02c28-f2bd-387c-9c8c-b2069e0e6159"
                    },
                    "title": {
                        "description": "Book title",
                        "type": "string",
                        "example": "Clean Arch"
                    },
                    "description": {
                        "description": "Book description",
                        "type": "string",
                        "example": "Book of Clean Arch"
                    },
                    "quantity": {
                        "description": "Book Quantity",
                        "type": "integer",
                        "example": "1"
                    },
                    "price": {
                        "description": "Book price",
                        "type": "integer",
                        "example": "100"
                    },
                    "author": {
                        "$ref": "#/components/schemas/ShowAuthorResource"
                    }
                },
                "type": "object"
            },
            "DefaultResponse": {
                "properties": {
                    "success": {
                        "description": "Indicates that the request was successful",
                        "type": "boolean"
                    },
                    "request": {
                        "description": "Requested Url",
                        "type": "string",
                        "example": "http://localhost/api/v1"
                    },
                    "method": {
                        "description": "Requested Method",
                        "type": "string",
                        "example": "Get"
                    },
                    "code": {
                        "description": "Status code",
                        "type": "string",
                        "example": "200"
                    },
                    "error": {
                        "description": "Error Message",
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "description": "Api data response",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}